feat: add ZELLIJ_DISABLE_AUTO_EXIT_ONLY_ONCE

+++ functions/zellij-auto-start.fish
@@ -1,6 +1,8 @@
 function zellij-auto-start
     set -q ZELLIJ_PANE_ID && return
 
+    set -U ZELLIJ_DISABLE_AUTO_EXIT_ONLY_ONCE false
+
     set current_number (zellij list-sessions | rg "automatic-(\w+)" -or '$1' | sort -rV | head -n 1)
     test -z "$current_number" && set current_number 0
     set new_number (math "$current_number" + 1)
@@ -9,7 +11,9 @@ function zellij-auto-start
         zellij -s "automatic-$new_number" || exit
 
         if $ZELLIJ_AUTO_EXIT
-            kill $fish_pid
+            if not $ZELLIJ_DISABLE_AUTO_EXIT_ONLY_ONCE
+                kill $fish_pid
+            end
         end
     end
 end
